code/requirements.txt
alabaster==0.7.10
alembic==0.9.5
Babel==2.5.0
bael.project==0.2.7
baelfire==0.5.1
Beaker==1.9.0
certifi==2017.7.27.1
chardet==3.0.4
coverage==4.4.1
docutils==0.14
fanstatic==1.0a5
fixdep==0.1
formskit==0.5.4.10
freezegun==0.3.9
future==0.16.0
Hamlish-Jinja==0.3.3
hupper==1.0
idna==2.6
imagesize==0.7.1
impaf==0.1.1
impaf-alembic==0.1
impaf-auth==0.1.1
impaf-beaker==0.1
impaf-fanstatic==0.1.1
impaf-flashmsg==0.1
impaf-formskit==0.1.3
impaf-haml==0.1
impaf-jinja2==0.1
impaf-sqlalchemy==0.1
Jinja2==2.9.6
js.bootstrap==3.3.4
js.jquery==1.9.1
Mako==1.0.7
MarkupSafe==1.0
mock==2.0.0
MorfDict==0.4.7
oauthlib==2.0.3
PasteDeploy==1.5.2
pbr==3.1.1
plaster==0.5
plaster-pastedeploy==0.4.1
psycopg2==2.7.3.1
py==1.4.34
Pygments==2.2.0
pyramid==1.9.1
pyramid-beaker==0.8
pyramid-debugtoolbar==4.3
pyramid-jinja2==2.7
pyramid-mako==1.0.2
pytest==3.2.2
pytest-cov==2.5.1
python-dateutil==2.6.1
python-editor==1.0.3
python-twitter==3.0rc1
pytz==2017.2
PyYAML==3.12
repoze.lru==0.7
requests==2.18.4
requests-oauthlib==0.8.0
shutilwhich==1.1.0
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
SQLAlchemy==0.9.9
translationstring==1.3
urllib3==1.22
uWSGI==2.0.15
venusian==1.1.0
virtualenv==15.1.0
waitress==1.0.2
WebOb==1.7.3
zope.deprecation==4.3.0
zope.interface==4.4.2
